Totally reworked lighting
Lighting now works based on a system where a lightable object has an on_lit and on_unlit method. The object counts how many lights are shining on it and determines if it is lit or not. Other objects may then check whether an object is lit by simply checking this value.